requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
azure-ai-formrecognizer==3.3.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.2.2
openai==1.11.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
